separate model validation from construction
